Refactor capybarize method.
