+++ lib/panel/capybarize/index.js
@@ -1,0 +1,5 @@
+import capybarize from './capybarize';
+
+export {
+  capybarize,
+};
